refactored to resolve name conflict
