+++ rank_llm/trec_eval.py
@@ -9,9 +9,9 @@ import tempfile
 from pyserini.search import get_qrels_file
 from pyserini.util import download_evaluation_script
 
-from pyserini_retriever import RetrievalMethod
-from rank_llm import PromptMode
-from topics_dict import TOPICS
+from rank_llm.pyserini_retriever import RetrievalMethod
+from rank_llm.rankllm import PromptMode
+from rank_llm.topics_dict import TOPICS
 
 
 class EvalFunction:
